refactor: Car, CarRacing, Utils를 다른 파일로 추출
- Car.java: 자동차 이름, 전진 등의 멤버를 담고 있는 클래스
- CarRacing.java: 자동차 경주를 진행하는 모든 메서드를 담고 있는 클래스(rafactoring 예정)
- Utils.java: class Utils에 속하는 코드(자동차/자동차 경주 클래스 둘 중 어디도 속하지 않는 코드)
